+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Image toolbox Changelog
 
+## 2.1.0 - 2023.05.15
+### Added
+* Added pictureMultiple() method, allowing for generating picture element with each source using different asset.
+* Added file placeholder mode, which generates placeholders based on source image file.
+* Transform layouts can now define HTML attributes using anonymous function. This allows defining attributes using asset object attributes.
+* Transform layouts and pictureMultiple() methods can use control panel defined image transforms.
+
+### Fixed
+* Fixed bug with placeholder not generating correctly when showing width and height is enabled.
+* Fixed bug with of throwing error whan trying to get width and height of svg file.
+
+### Deprecated
+* `picture()`, `pictureMedia()`, `pictureMax()` and `pictureMin()` are now deprecated.
+
 ## 2.0.2 - 2023.05.15
 ### Fixed
 * Width and height attributes are now properly calculated when width or height is missing from the transform.